Move mamut view classes in the respective views subpackages.
