+++ src/main/java/org/mastodon/mamut/WindowManager.java
@@ -54,15 +54,23 @@ import org.mastodon.feature.ui.FeatureColorModeConfigPage;
 import org.mastodon.mamut.feature.MamutFeatureProjectionsManager;
 import org.mastodon.mamut.model.Model;
 import org.mastodon.mamut.model.Spot;
+import org.mastodon.mamut.views.MamutView;
 import org.mastodon.mamut.views.MamutViewFactory;
 import org.mastodon.mamut.views.MamutViewI;
+import org.mastodon.mamut.views.bdv.MamutBranchViewBdv;
 import org.mastodon.mamut.views.bdv.MamutBranchViewBdvFactory;
+import org.mastodon.mamut.views.bdv.MamutViewBdv;
 import org.mastodon.mamut.views.bdv.MamutViewBdvFactory;
+import org.mastodon.mamut.views.grapher.MamutViewGrapher;
 import org.mastodon.mamut.views.grapher.MamutViewGrapherFactory;
 import org.mastodon.mamut.views.table.MamutViewSelectionTableFactory;
+import org.mastodon.mamut.views.table.MamutViewTable;
 import org.mastodon.mamut.views.table.MamutViewTableFactory;
+import org.mastodon.mamut.views.trackscheme.MamutBranchViewTrackScheme;
 import org.mastodon.mamut.views.trackscheme.MamutBranchViewTrackSchemeFactory;
+import org.mastodon.mamut.views.trackscheme.MamutBranchViewTrackSchemeHierarchy;
 import org.mastodon.mamut.views.trackscheme.MamutBranchViewTrackSchemeHierarchyFactory;
+import org.mastodon.mamut.views.trackscheme.MamutViewTrackScheme;
 import org.mastodon.mamut.views.trackscheme.MamutViewTrackSchemeFactory;
 import org.mastodon.model.tag.ui.TagSetDialog;
 import org.mastodon.ui.coloring.ColorBarOverlay.Position;
