+++ src/main/java/org/mastodon/mamut/io/MamutViewStateSerialization.java
@@ -28,15 +28,15 @@
  */
 package org.mastodon.mamut.io;
 
-import static org.mastodon.mamut.MamutBranchView.BRANCH_GRAPH;
-import static org.mastodon.mamut.MamutView.COLORBAR_POSITION_KEY;
-import static org.mastodon.mamut.MamutView.COLORBAR_VISIBLE_KEY;
-import static org.mastodon.mamut.MamutView.FEATURE_COLOR_MODE_KEY;
-import static org.mastodon.mamut.MamutView.FRAME_POSITION_KEY;
-import static org.mastodon.mamut.MamutView.GROUP_HANDLE_ID_KEY;
-import static org.mastodon.mamut.MamutView.NO_COLORING_KEY;
-import static org.mastodon.mamut.MamutView.SETTINGS_PANEL_VISIBLE_KEY;
-import static org.mastodon.mamut.MamutView.TAG_SET_KEY;
+import static org.mastodon.mamut.views.MamutBranchView.BRANCH_GRAPH;
+import static org.mastodon.mamut.views.MamutView.COLORBAR_POSITION_KEY;
+import static org.mastodon.mamut.views.MamutView.COLORBAR_VISIBLE_KEY;
+import static org.mastodon.mamut.views.MamutView.FEATURE_COLOR_MODE_KEY;
+import static org.mastodon.mamut.views.MamutView.FRAME_POSITION_KEY;
+import static org.mastodon.mamut.views.MamutView.GROUP_HANDLE_ID_KEY;
+import static org.mastodon.mamut.views.MamutView.NO_COLORING_KEY;
+import static org.mastodon.mamut.views.MamutView.SETTINGS_PANEL_VISIBLE_KEY;
+import static org.mastodon.mamut.views.MamutView.TAG_SET_KEY;
 import static org.mastodon.mamut.views.grapher.MamutViewGrapherFactory.GRAPHER_TRANSFORM_KEY;
 import static org.mastodon.mamut.views.table.MamutViewTableFactory.TABLE_DISPLAYED;
 import static org.mastodon.mamut.views.table.MamutViewTableFactory.TABLE_ELEMENT;
@@ -58,19 +58,19 @@ import java.util.Map.Entry;
 import javax.swing.JViewport;
 
 import org.jdom2.Element;
-import org.mastodon.mamut.MamutBranchViewBdv;
-import org.mastodon.mamut.MamutBranchViewTrackScheme;
-import org.mastodon.mamut.MamutView;
-import org.mastodon.mamut.MamutViewBdv;
-import org.mastodon.mamut.MamutViewGrapher;
-import org.mastodon.mamut.MamutViewTable;
-import org.mastodon.mamut.MamutViewTrackScheme;
 import org.mastodon.mamut.WindowManager;
 import org.mastodon.mamut.model.Link;
 import org.mastodon.mamut.model.Spot;
 import org.mastodon.mamut.model.branch.BranchLink;
 import org.mastodon.mamut.model.branch.BranchSpot;
+import org.mastodon.mamut.views.MamutView;
 import org.mastodon.mamut.views.MamutViewI;
+import org.mastodon.mamut.views.bdv.MamutBranchViewBdv;
+import org.mastodon.mamut.views.bdv.MamutViewBdv;
+import org.mastodon.mamut.views.grapher.MamutViewGrapher;
+import org.mastodon.mamut.views.table.MamutViewTable;
+import org.mastodon.mamut.views.trackscheme.MamutBranchViewTrackScheme;
+import org.mastodon.mamut.views.trackscheme.MamutViewTrackScheme;
 import org.mastodon.ui.coloring.ColorBarOverlay;
 import org.mastodon.ui.coloring.ColorBarOverlay.Position;
 import org.mastodon.ui.coloring.ColoringModel;
