+++ src/main/java/org/mastodon/mamut/views/grapher/MamutViewGrapherFactory.java
@@ -2,7 +2,6 @@ package org.mastodon.mamut.views.grapher;
 
 import java.util.Map;
 
-import org.mastodon.mamut.MamutViewGrapher;
 import org.mastodon.mamut.ProjectModel;
 import org.mastodon.mamut.views.AbstractMamutViewFactory;
 import org.mastodon.mamut.views.MamutViewFactory;
