+++ src/main/java/org/mastodon/mamut/views/trackscheme/MamutBranchViewTrackSchemeFactory.java
@@ -2,7 +2,6 @@ package org.mastodon.mamut.views.trackscheme;
 
 import java.util.Map;
 
-import org.mastodon.mamut.MamutBranchViewTrackScheme;
 import org.mastodon.mamut.ProjectModel;
 import org.mastodon.mamut.views.AbstractMamutViewFactory;
